+++ src/libraries/Common/tests/TestUtilities/System/IO/FileCleanupTestBase.cs
@@ -2,6 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using Xunit;
+using Microsoft.Win32.SafeHandles;
+using System.Buffers;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Threading;
@@ -77,6 +79,16 @@ namespace System.IO
         /// </summary>
         protected string TestDirectory { get; }
 
+        protected string GetRandomFileName() => GetTestFileName() + ".txt";
+        protected string GetRandomLinkName() => GetTestFileName() + ".link";
+        protected string GetRandomDirName()  => GetTestFileName() + "_dir";
+
+        protected string GetRandomFilePath() => Path.Join(ActualTestDirectory.Value, GetRandomFileName());
+        protected string GetRandomLinkPath() => Path.Join(ActualTestDirectory.Value, GetRandomLinkName());
+        protected string GetRandomDirPath()  => Path.Join(ActualTestDirectory.Value, GetRandomDirName());
+
+        private Lazy<string> ActualTestDirectory => new Lazy<string>(() => GetTestDirectoryActualCasing());
+
         /// <summary>Gets a test file full path that is associated with the call site.</summary>
         /// <param name="index">An optional index value to use as a suffix on the file name.  Typically a loop index.</param>
         /// <param name="memberName">The member name of the function calling this method.</param>
@@ -129,5 +141,56 @@ namespace System.IO
                 lineNumber,
                 index.GetValueOrDefault(),
                 Guid.NewGuid().ToString("N").Substring(0, 8)); // randomness to avoid collisions between derived test classes using same base method concurrently
+
+        private const int OPEN_EXISTING = 3;
+        private const int FILE_FLAG_BACKUP_SEMANTICS = 0x02000000;
+
+        // Some Windows versions like Windows Nano Server have the %TEMP% environment variable set to "C:\TEMP" but the
+        // actual folder name is "C:\Temp", which prevents asserting path values using Assert.Equal due to case sensitiveness.
+        // So instead of using TestDirectory directly, we retrieve the real path with proper casing of the initial folder path.
+        private unsafe string GetTestDirectoryActualCasing()
+        {
+            if (!PlatformDetection.IsWindows)
+                return TestDirectory;
+
+            try
+            {
+                using SafeFileHandle handle = Interop.Kernel32.CreateFile(
+                            TestDirectory,
+                            dwDesiredAccess: 0,
+                            dwShareMode: FileShare.ReadWrite | FileShare.Delete,
+                            dwCreationDisposition: FileMode.Open,
+                            dwFlagsAndAttributes:
+                                OPEN_EXISTING |
+                                FILE_FLAG_BACKUP_SEMANTICS // Necessary to obtain a handle to a directory
+                            );
+
+                if (!handle.IsInvalid)
+                {
+                    const int InitialBufferSize = 4096;
+                    char[]? buffer = ArrayPool<char>.Shared.Rent(InitialBufferSize);
+                    uint result = GetFinalPathNameByHandle(handle, buffer);
+
+                    // Remove extended prefix
+                    int skip = PathInternal.IsExtended(buffer) ? 4 : 0;
+
+                    return new string(
+                        buffer,
+                        skip,
+                        (int)result - skip);
+                }
+            }
+            catch { }
+
+            return TestDirectory;
+        }
+
+        private unsafe uint GetFinalPathNameByHandle(SafeFileHandle handle, char[] buffer)
+        {
+            fixed (char* bufPtr = buffer)
+            {
+                return Interop.Kernel32.GetFinalPathNameByHandle(handle, bufPtr, (uint)buffer.Length, Interop.Kernel32.FILE_NAME_NORMALIZED);
+            }
+        }
     }
 }
