+++ src/libraries/System.Diagnostics.FileVersionInfo/tests/System.Diagnostics.FileVersionInfo.Tests/FileVersionInfoTest.Unix.cs
@@ -23,7 +23,7 @@ namespace System.Diagnostics.Tests
         public void Symlink_ValidFile_Succeeds()
         {
             string filePath = Path.Combine(Directory.GetCurrentDirectory(), TestAssemblyFileName);
-            string linkPath = Path.ChangeExtension(GetTestFilePath(), ".link");
+            string linkPath = GetRandomLinkPath();
 
             Assert.Equal(0, symlink(filePath, linkPath));
 
@@ -68,7 +68,7 @@ namespace System.Diagnostics.Tests
             string sourcePath = Path.Combine(Directory.GetCurrentDirectory(), TestAssemblyFileName);
             string filePath = GetTestFilePath();
             File.Copy(sourcePath, filePath);
-            string linkPath = Path.ChangeExtension(GetTestFilePath(), ".link");
+            string linkPath = GetRandomLinkPath();
             Assert.Equal(0, symlink(filePath, linkPath));
             File.Delete(filePath);
             Assert.Throws<FileNotFoundException>(() => FileVersionInfo.GetVersionInfo(linkPath));
