+++ src/libraries/System.IO.FileSystem.Watcher/tests/FileSystemWatcher.File.Delete.cs
@@ -98,7 +98,7 @@ namespace System.IO.Tests
             using (var watcher = new FileSystemWatcher(dir.Path, "*"))
             {
                 // Make the symlink in our path (to the temp file) and make sure an event is raised
-                string symLinkPath = Path.Combine(dir.Path, Path.ChangeExtension(Path.GetFileName(temp.Path), ".link"));
+                string symLinkPath = Path.Combine(dir.Path, GetRandomLinkName());
                 Action action = () => File.Delete(symLinkPath);
                 Action cleanup = () => Assert.True(MountHelper.CreateSymbolicLink(symLinkPath, temp.Path, false));
                 cleanup();
