+++ src/libraries/System.IO.FileSystem/tests/Base/SymbolicLinks/BaseSymbolicLinks.Unix.cs
@@ -14,6 +14,5 @@ namespace System.IO.Tests
 {
     public abstract partial class BaseSymbolicLinks : FileSystemTest
     {
-        private string GetTestDirectoryActualCasing() => TestDirectory;
     }
 }
