+++ src/libraries/System.IO.FileSystem/tests/Base/SymbolicLinks/BaseSymbolicLinks.cs
@@ -29,16 +29,6 @@ namespace System.IO.Tests
             return (DirectoryInfo)Directory.CreateSymbolicLink(path, path);
         }
 
-        protected string GetRandomFileName() => GetTestFileName() + ".txt";
-        protected string GetRandomLinkName() => GetTestFileName() + ".link";
-        protected string GetRandomDirName()  => GetTestFileName() + "_dir";
-
-        protected string GetRandomFilePath() => Path.Join(ActualTestDirectory.Value, GetRandomFileName());
-        protected string GetRandomLinkPath() => Path.Join(ActualTestDirectory.Value, GetRandomLinkName());
-        protected string GetRandomDirPath()  => Path.Join(ActualTestDirectory.Value, GetRandomDirName());
-
-        private Lazy<string> ActualTestDirectory => new Lazy<string>(() => GetTestDirectoryActualCasing());
-
         /// <summary>
         /// Changes the current working directory path to a new temporary directory.
         /// Important: Make sure to call this inside a remote executor to avoid changing the cwd for all tests in same process.
