+++ src/libraries/System.IO.FileSystem/tests/Directory/SetCurrentDirectory.cs
@@ -53,7 +53,7 @@ namespace System.IO.Tests
                 RemoteExecutor.Invoke(() =>
                 {
                     var path = GetTestFilePath();
-                    var linkPath = Path.ChangeExtension(GetTestFilePath(), ".link");
+                    var linkPath = GetRandomLinkPath();
 
                     Directory.CreateDirectory(path);
                     Assert.True(MountHelper.CreateSymbolicLink(linkPath, path, isDirectory: true));
