+++ src/libraries/System.IO.FileSystem/tests/FileInfo/Exists.cs
@@ -104,7 +104,7 @@ namespace System.IO.Tests
         public void SymLinksMayExistIndependentlyOfTarget()
         {
             var path = GetTestFilePath();
-            var linkPath = Path.ChangeExtension(GetTestFilePath(), ".link");
+            var linkPath = GetRandomLinkPath();
 
             var pathFI = new FileInfo(path);
             var linkPathFI = new FileInfo(linkPath);
